Fix model functions using SerializeIter
