+++ src/http/client.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::missing_errors_doc)]
 
 use std::borrow::Cow;
+use std::cell::Cell;
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
 
@@ -13,6 +14,7 @@ use reqwest::Url;
 use reqwest::{Client, ClientBuilder, Response as ReqwestResponse, StatusCode};
 use secrecy::{ExposeSecret as _, Secret};
 use serde::de::DeserializeOwned;
+use serde::ser::SerializeSeq as _;
 use serde_json::{from_value, json, to_string, to_vec};
 use tracing::{debug, trace};
 
@@ -64,6 +66,29 @@ impl secrecy::Zeroize for Token {
 
 impl secrecy::CloneableSecret for Token {}
 impl secrecy::DebugSecret for Token {}
+
+// NOTE: This cannot be passed in from outside, due to `Cell` being !Send.
+struct SerializeIter<I>(Cell<Option<I>>);
+
+impl<I> SerializeIter<I> {
+    pub fn new(iter: I) -> Self {
+        Self(Cell::new(Some(iter)))
+    }
+}
+
+impl<Iter, Item> serde::Serialize for SerializeIter<Iter>
+where
+    Iter: Iterator<Item = Item>,
+    Item: serde::Serialize,
+{
+    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
+        let Some(iter) = self.0.take() else {
+            return serializer.serialize_seq(Some(0))?.end();
+        };
+
+        serializer.collect_seq(iter)
+    }
+}
 
 /// A builder for the underlying [`Http`] client that performs requests to Discord's HTTP API. If
 /// you do not need to use a proxy or do not need to disable the rate limiter, you can use
@@ -1671,9 +1696,9 @@ impl Http {
     pub async fn edit_guild_channel_positions(
         &self,
         guild_id: GuildId,
-        value: &impl serde::Serialize,
+        value: impl Iterator<Item: serde::Serialize>,
     ) -> Result<()> {
-        let body = to_vec(value)?;
+        let body = to_vec(&SerializeIter::new(value))?;
 
         self.wind(204, Request {
             body: Some(body),
@@ -2010,12 +2035,14 @@ impl Http {
     pub async fn edit_role_positions(
         &self,
         guild_id: GuildId,
-        map: &impl serde::Serialize,
+        positions: impl Iterator<Item: serde::Serialize>,
         audit_log_reason: Option<&str>,
     ) -> Result<Vec<Role>> {
+        let body = to_vec(&SerializeIter::new(positions))?;
+
         let mut value: Value = self
             .fire(Request {
-                body: Some(to_vec(&map)?),
+                body: Some(body),
                 multipart: None,
                 headers: audit_log_reason.map(reason_into_header),
                 method: LightMethod::Patch,
