+++ src/model/utils.rs
@@ -1,9 +1,7 @@
-use std::cell::Cell;
 use std::fmt;
 
 use arrayvec::ArrayVec;
 use serde::de::Error as DeError;
-use serde::ser::SerializeSeq;
 use serde_cow::CowStr;
 use small_fixed_array::FixedString;
 
@@ -68,28 +66,6 @@ where
     E: serde::de::Error,
 {
     remove_from_map_opt(map, key)?.ok_or_else(|| serde::de::Error::missing_field(key))
-}
-
-pub(super) struct SerializeIter<I>(Cell<Option<I>>);
-
-impl<I> SerializeIter<I> {
-    pub fn new(iter: I) -> Self {
-        Self(Cell::new(Some(iter)))
-    }
-}
-
-impl<Iter, Item> serde::Serialize for SerializeIter<Iter>
-where
-    Iter: Iterator<Item = Item>,
-    Item: serde::Serialize,
-{
-    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let Some(iter) = self.0.take() else {
-            return serializer.serialize_seq(Some(0))?.end();
-        };
-
-        serializer.collect_seq(iter)
-    }
 }
 
 pub(super) enum StrOrInt<'de> {
